feat: Improve type portability and add doorbell callback
Refactor `zbc_semihost` parameters and return type to use `uintptr_t` for better portability across different architectures, aligning with picolibc's `sys_semihost` expectations.

- Changed `zbc_semihost` return type from `long` to `uintptr_t`.
- Changed `op` and `param` types in `zbc_semihost` from `unsigned long` and `void*` to `uintptr_t`.
- Updated `zbc_response.result` from `long` to `int`.
- Modified `zbc_builder_add_parm_int` and `zbc_builder_add_parm_uint` to use `int` and `unsigned int` respectively, promoting `int` as the standard integer type.
- Added `doorbell_callback` and `doorbell_ctx` to `zbc_client_state_t` to provide a test hook for doorbell triggers.
- Updated documentation to reflect the new type recommendations, specifically advising against `long` due to portability concerns and suggesting `unsigned int` over `uint32_t` where possible.

+++ include/zbc_semi_host.h
@@ -187,9 +187,9 @@ int zbc_host_process(zbc_host_state_t *state, uint64_t riff_addr);
  * state: Host state (for endianness and int_size info)
  * buf: Buffer containing value (int_size bytes)
  *
- * Returns: Integer value (sign-extended to long)
+ * Returns: Integer value (sign-extended to int)
  */
-long zbc_host_read_int(const zbc_host_state_t *state, const unsigned char *buf);
+int zbc_host_read_int(const zbc_host_state_t *state, const unsigned char *buf);
 
 /*
  * Write an integer to guest-endian buffer.
@@ -199,7 +199,7 @@ long zbc_host_read_int(const zbc_host_state_t *state, const unsigned char *buf);
  * value: Value to write
  */
 void zbc_host_write_int(const zbc_host_state_t *state, unsigned char *buf,
-                        long value);
+                        int value);
 
 /*
  * Read a pointer from guest-endian buffer.
